* GiftVoucherType.java: byte -> int; + EdmTypeKind.java; #15

+++ src/main/java/io/githhub/TradersTeam/grand_node_java_client/enums/GiftVoucherType.java
@@ -1,11 +1,11 @@
 package io.githhub.TradersTeam.grand_node_java_client.enums;
 
 public enum GiftVoucherType {
-    Virtual((byte) 0),
-    Physical((byte) 1);
-    byte value;
+    Virtual(0),
+    Physical(1);
+    int value;
 
-    GiftVoucherType(byte value) {
+    GiftVoucherType(int value) {
         this.value = value;
     }
 }
